add dns to qemu environment
Qemu is used to run the apt-get target executable, but it can fail if
DNS lookups are required of the APT repository hosts or the $http_proxy
server. The comments talk about both /etc/hosts and DNS, but the code
doesn't do anything about the latter. Copy /etc/resolv.conf from the
host environment into the chroot jail.

Signed-off-by: David Wolfe <david.wolfe@nxp.com>