Fix MGS-6634 [FT][Network]: unable to browse internet after boot (DHCP issue)
David Wolfe added a comment - 05/May/22 20:59
This and MGS-6636 appear to be the same issue. MGS-6636 is definitely a DNS issue. The solution I found was:

sudo apt remove connman
sudo systemctl enable systemd-resolved

(It may then be necessary to reboot or at least bounce the network connection.) Network connectivity worked perfectly for me after this one change, even Internet browsing with Chromium. It works both inside the NXP WAN and outside. It works over both WiFi and 802.11 Ethernet.

Connman and Network Manager perform the same functionality (DHCP + DNS configuration provided by the DHCP lease) and, therefore, conflict. They should not both be installed and used to manage the same network interfaces. The GNOME desktop interacts with Network Manager, so it should be the tool chosen for the Desktop PoC (just like arch: amd64 installs).
Network Manager works with dnsmasq, systemd-resolved or some other sophisticated caching DNS server on the local machine to maintain correct name lookups for different interfaces. Ubuntu has come to rely more on systemd over the years, so that is the correct choice for 20.04.
Network Manager as configured correctly obtains a DHCP lease (look at ip addr or ifconfig). This lease includes information about the upstream DNS servers on this network. However, Connman takes over UDP/53 (DNS server) even though it has no way of getting that list of DNS servers from Network Manager. This is the conflict.

Launching udhcpc(8) by hand solves the problem because it changes /etc/resolv.conf with the old-style assumption that there is one domain for all combined network interfaces. Editing /etc/resolv.conf by hand or via udhcpc is 1) a limited assumption and 2) fails to integrate into the GNOME desktop settings GUI program(s) which is kind of the point of this PoC.

Signed-off-by: Xianzhong Li <xianzhong.li@nxp.com>